requirements.txt
numpy==1.23.3
Pillow==9.4.0
requests==2.28.1
unicode_slugify==0.1.5
